Drop Elasticsearch5SearchBackend and make Elasticsearch6SearchBackend the base implementation
